Make span context (re)construction more ergonomic
This commit allows one to continue a trace asynchronously, from one
process to another, by only requiring the use of the `trace_id` and the
`span_id` of the initial trace.

The additional implementation of `set_current_span/2` in `:otel_tracer`, allows
a span context to be (re)constructed using a `trace_id` and a `span_id`.

This commit also implements the functionality of the `:context` start option,
which is used to optionally set the current context.

Additionally, it implements the functionality of the `:parent` start option so
that `parent: {trace_id, span_id}` can be used to set the parent span.